+++ flutter_inappwebview_forge_macos/macos/flutter_inappwebview_forge_macos/Sources/flutter_inappwebview_forge_macos/InAppWebView/WebViewChannelDelegate.swift
@@ -173,6 +173,10 @@ public class WebViewChannelDelegate: ChannelDelegate {
             }
             result(true)
             break
+        case .setContextMenu:
+            webView?.contextMenu = arguments?["contextMenu"] as? [String: Any?]
+            result(true)
+            break
         case .getSettings:
             if let iabController = webView?.inAppBrowserDelegate as? InAppBrowserWebViewController {
                 result(iabController.getSettings())