+++ flutter_inappwebview_forge_macos/macos/flutter_inappwebview_forge_macos/Sources/flutter_inappwebview_forge_macos/InAppWebView/WebViewChannelDelegateMethods.swift
@@ -30,6 +30,7 @@ public enum WebViewChannelDelegateMethods: String {
     case isLoading = "isLoading"
     case takeScreenshot = "takeScreenshot"
     case setSettings = "setSettings"
+    case setContextMenu = "setContextMenu"
     case getSettings = "getSettings"
     case close = "close"
     case show = "show"